Eina: Fix clang warnings
Use fabs for double values

+++ src/lib/eina/eina_matrix.c
@@ -65,13 +65,13 @@ static inline double
    const double B = 4/M_PI;
    const double C = -4/(M_PI*M_PI);
 
-   double y = (B * x) + (C * x * fabsf(x));
+   double y = (B * x) + (C * x * fabs(x));
 
 #ifdef EXTRA_PRECISION
    //  const float Q = 0.775;
    const double P = 0.225;
 
-   y = P * (y * fabsf(y) - y) + y; // Q * y + P * y * abs(y)
+   y = P * (y * fabs(y) - y) + y; // Q * y + P * y * abs(y)
 #endif
    return y;
 }
